+++ src/aks-preview/azext_aks_preview/_help.py
@@ -361,7 +361,7 @@ helps['aks create'] = """
           short-summary: (PREVIEW) Enable pod identity addon for cluster using Kubnet network plugin.
         - name: --enable-workload-identity
           type: bool
-          short-summary: (PREVIEW) Enable workload identity addon.
+          short-summary: Enable workload identity addon.
         - name: --disable-disk-driver
           type: bool
           short-summary: Disable AzureDisk CSI Driver.
@@ -650,10 +650,7 @@ helps['aks update'] = """
           short-summary: (PREVIEW) Disable Pod Identity addon for cluster.
         - name: --enable-workload-identity
           type: bool
-          short-summary: (PREVIEW) Enable Workload Identity addon for cluster.
-        - name: --disable-workload-identity
-          type: bool
-          short-summary: (PREVIEW) Disable Workload Identity addon for cluster.
+          short-summary: Enable Workload Identity addon for cluster.
         - name: --enable-secret-rotation
           type: bool
           short-summary: Enable secret rotation. Use with azure-keyvault-secrets-provider addon.